Move plugins/adapters back to ROM::Global
